Feature extraction script for Respiration, Heart rate, GSR, and temperature
At the end of the script the features are all stored in feats array, the name of each column in feats is in the name variable and all_errors is a dict containing trials that do not provide good signals. The key corresponds to the subject number and then there is an array of each trial for which at least one signal was invalid.